landings_apis/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User


class UserCreateSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ['username', 'password', 'first_name', 'last_name', 'email']

    def create(self, validated_data):
        username = validated_data['username']
        password = validated_data['password']
        firstname = validated_data['first_name']
        lastname = validated_data['last_name']
        email = validated_data['email']

        new_user = User(username=username,
                        first_name=firstname, last_name=lastname, email=email)
        new_user.set_password(password)
        new_user.save()
        return validated_data
